// src/app/api/models/item-selector-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ItemSelectorResponseDto {
  category?: string;
  description?: string;

  /**
   * Duration in minutes for services
   */
  duration?: number;
  id: number;
  name: string;
  price: number;
  status: string;
  thumbnailUrl?: string;
  type: 'product' | 'service';
}